Style: new style for template-add dialog
