+++ libs/material/src/lib/template/template-item/template-item.component.html
@@ -1,5 +1,5 @@
 <mat-card class="mat-elevation-z8" fxLayout="row" fxLayoutAlign="start center">
-  <div fxLayout="column" fxLayoutAlign="center start">
+  <div class="text-block" fxLayout="column" fxLayoutAlign="center start">
     <mat-card-title fxFlex="50">
       {{ template.name }}
     </mat-card-title>
