+++ libs/material/src/lib/template/template-item/template-item.component.scss
@@ -3,6 +3,11 @@ mat-card-title {
   font-size: 20px;
   line-height: 24px;
   letter-spacing: 0.25px;
+  margin-bottom: 0;
+}
+
+.text-block {
+  margin-bottom: -5px;
 }
 
 .modification-text {
@@ -30,5 +35,6 @@ mat-card-title {
 
 mat-card {
   margin: 5% 5%;
-  min-height: 100px;
+  min-height: 90px;
+  border-radius: 0px;
 }
